+++ src/components/LocationAutocompleteInput/GeocoderGoogleMaps.js
@@ -1,0 +1,69 @@
+import React from 'react';
+import { getPlacePredictions, getPlaceDetails } from '../../util/googleMaps';
+
+import css from './LocationAutocompleteInput.css';
+
+// When displaying data from the Google Maps Places API, and
+// attribution is required next to the results.
+// See: https://developers.google.com/places/web-service/policies#powered
+export const GeocoderAttribution = () => <div className={css.poweredByGoogle} />;
+
+/**
+ * A forward geocoding (place name -> coordinates) implementation
+ * using the Google Maps Places API.
+ */
+class GeocoderGoogleMaps {
+  constructor() {
+    this.sessionToken = null;
+  }
+  getSessionToken() {
+    this.sessionToken =
+      this.sessionToken || new window.google.maps.places.AutocompleteSessionToken();
+    return this.sessionToken;
+  }
+
+  // Public API
+  //
+
+  /**
+   * Search places with the given name.
+   *
+   * @param {String} search query for place names
+   *
+   * @return {Promise<{ search: String, predictions: Array<Object>}>}
+   * results of the geocoding, should have the original search query
+   * and an array of predictions. The format of the predictions is
+   * only relevant for the `getPlaceDetails` function below.
+   */
+  getPlacePredictions(search) {
+    return getPlacePredictions(search, this.getSessionToken()).then(results => {
+      return {
+        search,
+        predictions: results.predictions,
+      };
+    });
+  }
+
+  /**
+   * Get the address text of the given prediction.
+   */
+  getPredictionAddress(prediction) {
+    return prediction.description;
+  }
+
+  /**
+   * Fetch or read place details from the selected prediction.
+   *
+   * @param {Object} prediction selected prediction object
+   *
+   * @return {Promise<util.propTypes.place>} a place object
+   */
+  getPlaceDetails(prediction) {
+    return getPlaceDetails(prediction.place_id, this.getSessionToken()).then(place => {
+      this.sessionToken = null;
+      return place;
+    });
+  }
+}
+
+export default GeocoderGoogleMaps;