+++ src/util/types.js
@@ -72,7 +72,6 @@ propTypes.place = shape({
   address: string.isRequired,
   origin: propTypes.latlng,
   bounds: propTypes.latlngBounds, // optional viewport bounds
-  country: string, // country code, e.g. FI, US
 });
 
 // Denormalised image object